+++ style_generator/lib/src/data/variable_handler.dart
@@ -15,8 +15,13 @@ class VariableHandler {
   final ClassElement clazz;
   final ConstructorElement constructor;
 
-  /// map constructor elements and their declaration for [_lookupField]
+  /// Map constructor elements and their declaration for [_lookupField]
+  /// This one holds concrete implementations like
+  /// `GenParent<DataStuff, String>({required String some, required DataStuff? something})`
   final Map<ConstructorElement, ConstructorDeclaration> _constructors = {};
+  /// This one holds generic implementations like
+  /// `GenParent<T, L>({required String some, required DataStuff? something})`
+  final Map<ConstructorElement, ConstructorDeclaration> _baseConstructors = {};
 
   late final List<Variable> constructorParams;
   late final List<Variable> fields;
@@ -35,10 +40,10 @@ class VariableHandler {
   /// returns the corresponding list of variables by [type] to allow dynamic access
   @protected
   List<Variable> getVariablesByType(FieldType type) => switch (type) {
-    FieldType.FIELD => fields,
-    FieldType.CONSTRUCTOR_PARAM => constructorParams,
-    FieldType.MERGED => merged,
-  };
+        FieldType.FIELD => fields,
+        FieldType.CONSTRUCTOR_PARAM => constructorParams,
+        FieldType.MERGED => merged,
+      };
 
   VariableHandler({required this.clazz, required this.constructor}) {
     constructorParams = constructor.formalParameters.map((e) => Variable(element: e)).toList();
@@ -47,11 +52,58 @@ class VariableHandler {
 
   FieldElement _lookupField(Variable parameter) {
     // short circuit lookup if the field is already assigned
-    if (parameter.fieldElement != null) return parameter.fieldElement!;
+    if (parameter.fieldElement != null) parameter.fieldElement!;
 
     return _lookupTree(parameter.element);
   }
 
+
+
+  /// Will find the field of a constructor parameter.
+  ///
+  /// The following code shows a specific case were the type information of the constructor parameters are lost.
+  /// By default, if the inheritance structure wouldn't contain generic definitions, all SuperFormal..
+  /// would map very likely map to a FieldFormal.., but with the use of _multiple_ generic ancestors, they are all resolved
+  /// to FormalParameter.. which have a more complex lookup.
+  /// Example:
+  /// ```dart
+  /// class GenChild extends GenIntermediate<DataStuff, String> with _$GenChild {
+  ///   final String local;
+  ///
+  ///   const GenChild({            //     (underscore for better readability)
+  ///     required this.local,      // <-- Field_FormalParameterElement
+  ///     required super.some,      // <-- Super_FormalParameterElement
+  ///     required super.something, // <-- Super_FormalParameterElement
+  ///     required super.bim        // <-- Super_FormalParameterElement
+  ///   });
+  /// }
+  ///
+  /// class GenIntermediate<T, L> extends GenParent<T, L> {
+  ///   // all parameter here are SuperFormalParameterElements
+  ///   const GenIntermediate({required super.some, required super.something, required super.bim});
+  /// }
+  ///
+  /// class GenParent<T, L> {
+  ///   final String some;
+  ///   final T something;
+  ///   final L? bimbo;
+  ///   final String nothing;
+  ///
+  ///   const GenParent({
+  ///     required this.some,       // <-- without generics, this would be a FieldFormal.. but with generics, it is a Formal..
+  ///     required this.something,  // <-- without generics, this would be a FieldFormal.. but with generics, it is a Formal..
+  ///     required L? bim           // <-- this one is the _second_! special case
+  ///   }) :
+  ///         bimbo = bim,
+  ///         nothing = "";
+  /// }
+  /// ```
+  /// Important info: The returned [FieldElement] may be resolved with [ResolvedType] to something which
+  /// contains a [DartType] of [TypeParameterType].
+  /// This [TypeParameterType] represents a generic (like T, L, ...) and is mostly not what we need, because the [ResolvedType]
+  /// exists to actually resolve generics to a concrete class.
+  /// However: It is technically correct that this method returns the [FieldElement] of that type.
+  /// Therefor the [VariableHandler] will take care of the conversion.
   FieldElement _lookupTree(VariableElement element) {
     FieldElement? field;
 
@@ -61,16 +113,53 @@ class VariableHandler {
       case FieldFormalParameterElement():
         field = element.field;
       case SuperFormalParameterElement():
-        field = _lookupTree(element.superConstructorParameter!);
+        // [element.superConstructorParameter] looses type information
+        // (the returned type is always [FormalParameterElement], even for [FieldFormalParameterElement]s for example.
+        var superElement = element.superConstructorParameter;
+
+        field = _lookupTree(superElement!);
       case FormalParameterElement():
         var constructorElement = element.enclosingElement;
         ConstructorDeclaration? d = _constructors[constructorElement];
 
-        if (d != null) {
-          var map = d.mapParameterToField(lookup: (superParameter) => _lookupTree(superParameter).displayName);
+        FieldElement? checkInitializersOf(ConstructorDeclaration d) {
+          var map = d.mapInitializersToField(lookup: (superParameter) => _lookupTree(superParameter).displayName);
 
-          String fieldName = map[element.displayName]!;
-          field = fields.firstWhereOrNull((f) => f.displayName == fieldName)?.fieldElement;
+          String? fieldName = map[element.displayName];
+          return fields.firstWhereOrNull((f) => f.displayName == fieldName)?.fieldElement;
+        }
+
+        if (d != null) {
+          // We enter this part, when we have to map a constructor parameter via the initializers (the : behind the constructor)
+          // to the field.
+          field = checkInitializersOf(d);
+
+          assert(field != null, "We expected that parameter:'$element' must be mapped through initializers, but none were found");
+        } else {
+          // We enter here, when the Example from the doc above hits and we have to map the given constructor parameter
+          // 'element' like 'required DataStuff? something' to the constructor parameter 'required T? something'
+          d = _baseConstructors[constructorElement];
+
+          if (d != null) {
+            // Taken the example above, this one will find nothing for 'required DataStuff? something' and 'required String bim'.
+            // read the comments down below, this one is matched for the 'required L? bim' lookup (the generic).
+            field = checkInitializersOf(d);
+
+            // This one resolves
+            // - 'required DataStuff? something' to 'required this.something'
+            // - 'required String bim'           to 'required L? bim'
+            if (field == null) {
+              for (var param in d.parameters.parameters) {
+                if (param.declaredFragment == element.firstFragment) {
+                  // it then resolves
+                  // - 'requires this.something' to the field via the 'FieldFormalParameterElement'
+                  // - 'requires L? bim' enters again the 'FormalParameterElement' but hits the `field = checkInitializersOf(d);` above
+                  field = _lookupTree(param.declaredFragment!.element);
+                  break;
+                }
+              }
+            }
+          }
         }
     }
 
@@ -83,19 +172,23 @@ class VariableHandler {
 
     _constructors.clear();
     _constructors[constructor] = d;
+    _baseConstructors[constructor.baseElement] = d;
 
     ConstructorElement? superConstructor = constructor.superConstructor;
 
+    // @formatter:off
     while (superConstructor != null && (!superConstructor.library.isDartCore && !superConstructor.library.isDartAsync)) {
       ConstructorDeclaration? superDeclaration = await resolver.astNodeFor(superConstructor.firstFragment, resolve: true) as ConstructorDeclaration?;
 
       if (superDeclaration != null) {
         _constructors[superConstructor] = superDeclaration;
+        _baseConstructors[superConstructor.baseElement] = superDeclaration;
       }
 
       superConstructor = superConstructor.superConstructor;
     }
   }
+  // @formatter:on
 
   /// builds the annotation cache for [merged]
   void build<T>(AnnotationConverter<T> converter, {FieldType? type, bool? annotationTypeCheck}) {
@@ -108,7 +201,8 @@ class VariableHandler {
       anno = _findAnnotationForType<T>(type, v, converter);
       if (anno != null) {
         if (annotationTypeCheck) {
-          bool hasMatchingType = v.element.isOfSameTypeAsTypeArgumentFromObject(anno.object, lessStrict: true, allowDynamic: true);
+          bool hasMatchingType =
+              v.element.isOfSameTypeAsTypeArgumentFromObject(anno.object, lessStrict: true, allowDynamic: true);
           if (!hasMatchingType) styleKeyTypeMismatch(v, anno.object.type);
         }
 
@@ -116,7 +210,6 @@ class VariableHandler {
       }
     }
   }
-
 
   void resolveTypes(ResolvedLibraryResult resolvedLib, {FieldType? type}) {
     type ??= FieldType.MERGED;
@@ -130,7 +223,6 @@ class VariableHandler {
   @protected
   AnnotatedElement<T>? _findAnnotationForType<T>(FieldType type, Variable v, AnnotationConverter<T> converter) {
     switch (type) {
-
       case FieldType.FIELD:
         return _getAnnotation(null, v, converter);
       case FieldType.CONSTRUCTOR_PARAM: